refactor: selectbox 네이밍 select로 변경
